pyproject.toml
[project]
name = "ccr-cot-experiment"
version = "0.1.0"
description = "Counterfactual Consistency Re-ranking for Chain-of-Thought"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "datasets>=2.14.0",
    "openai>=1.12.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
]